[install] split out the install methods to an install-utils file
